Flag Update
Introduced flags to all story related tags plus a few minor ones, update the age bonus icons to reflect this, changed the icon of the remina government reform, changed the name of sister lamia to her canon name "Sophia" and increased her diplo stat, gave to her and the rulers of Succubus and Slug nuns the "Zealot" personality to reflect their canon counterparts. Corrected the typo in Twinkell's name for her nation.