Mission icons background
Added backgrounds to all non-AI icons that had transparency, fixed some errors with the size of devil_hunter_mission, and an icon for yamata no orochi whose faceset had a piece of another faceset attached to the bottom. Re-done my two mission icons from the Slug Queen MT on the irredentist missions on the Navy and Krakken as they look very rough compared to the rest. worked on the AI Alraune and doll generic mission icons but i will wait for approval for those.